test: guard all base import bans against biome override drift
Generalize the override-drift guard from @aws-sdk-only to the entire
noRestrictedImports base superset, so a NEW base ban that is not mirrored
into overrides[0] breaks the test on its own — no one has to remember to
extend the guard for each ban.

A Biome override replaces (does not merge) the rule's options, so any base
ban absent from overrides[0].paths silently stops applying to **/*.test.ts
and the auth-boundary files. The test now iterates every base key and asserts
it is re-declared verbatim in the override, except keys deliberately withheld
(@packages/domain/user, where authenticatedUserIdFrom is intentionally allowed
in test/auth-boundary files). That exception list is itself guarded: each
withheld key must still exist in the base, so it cannot rot silently.

Renamed biome-aws-sdk-ban-override-guard.test.ts ->
biome-restricted-imports-override-guard.test.ts to match the broader scope.
